requirements.txt
fastapi
uvicorn[standard]
bcrypt
python-jose[cryptography]
cryptography
python-multipart
sqlalchemy[asyncio]
aiosqlite
alembic
pydantic-settings
python-dotenv
slowapi